Add files via upload
We performed transmission simulations to compare BPSK, MSK, and QPSK modulation formats considering an AWGN channel. In the first step, we generated random data based on the aforementioned formats. After passing through an AWGN channel, the transmitted signals are received, and the bit error rate (BER) as a function of signal to noise ratio (SNR) is calculated for each modulation format. In the appendix, the report file of this project is attached.